make carrierManagemanet compatible with
This commit should make the carrier management compatible with common.
I dont understand why we have Carrier.java class in common package! This could be due to I was not able to attend meetings.
