+++ src/main/java/carrierManagementSystem/CarrierManagementIf.java
@@ -1,6 +1,9 @@
 package carrierManagementSystem;
 import java.util.List;
 
+import common.FrequencyBand;
+import common.RfPorts;
+
 public interface CarrierManagementIf {
     /**
      * The below method is responsible for creating LTE carrier based on inputs.
