+++ src/main/java/carrierManagementSystem/LteCarrierBuilder.java
@@ -2,6 +2,8 @@ package carrierManagementSystem;
 
 import java.util.ArrayList;
 import java.util.List;
+import common.FrequencyBand;
+import common.RfPorts;
 
 /**
  * This class is a child class to CarrierBuilder. It builds LTECarrier objects.
