+++ src/main/java/carrierManagementSystem/WcdmaCarrier.java
@@ -1,6 +1,8 @@
 package carrierManagementSystem;
 
 import java.util.List;
+import common.FrequencyBand;
+import common.RfPorts;
 
 /**
  * This class builds WCDMA carriers. The WCDMACarrierBuilder class uses this
